[dev-task9] - fragment_weather upper design added, language interceptor added, new fields added to domain weather class

+++ app/src/main/java/dev/gladkowski/wetaherapp/entity/weather/data/WindResponse.java
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName;
 public class WindResponse {
 
     @SerializedName("speed")
-    private Float speed;
+    private Double speed;
 
-    public Float getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 }
